My junkfood rides over the ocean

+++ code/module/JfaRest/src/JfaRest/Controller/JunkFoodRestController.php
@@ -4,8 +4,6 @@ namespace JfaRest\Controller;
 use Zend\Mvc\Controller\AbstractRestfulController;
 
 use Jfa\Model\JunkFood;
-use Jfa\Form\JunkFoodForm;
-use Jfa\Model\JunkFoodTable;
 use Zend\View\Model\JsonModel;
 
 class JunkFoodRestController extends AbstractRestfulController
@@ -13,6 +11,7 @@ class JunkFoodRestController extends AbstractRestfulController
     protected $junkFoodTable;
     protected $ingredientTable;
     protected $userTable;
+    protected $relationTable;
 
     protected $userName = false;
     protected $authservice;
@@ -56,7 +55,11 @@ class JunkFoodRestController extends AbstractRestfulController
 
     public function get($id)
     {
-        $data = $this->getJunkFoodTable()->getJunkFood($id);
+        if ($id == 'drogo') {
+            $data = $this->getJunkFoodTable()->getDrogoSugestion();
+        } else {
+            $data = $this->getJunkFoodTable()->getJunkFood($id);
+        }
 
         $result = new JsonModel(array('data' => $data));
         return $result;
@@ -69,6 +72,11 @@ class JunkFoodRestController extends AbstractRestfulController
         if ($this->getAuthService()->hasIdentity()){
             $this->userName = $this->getAuthService()->getIdentity();
             $userId = $this->getUserTable()->getUserByName($this->userName);
+        } else {
+            return new JsonModel(array(
+                'status' => 'failure',
+                'message' => 'You have to log in first'
+            ));
         }
 
         $data['userID'] = $userId;
@@ -77,28 +85,90 @@ class JunkFoodRestController extends AbstractRestfulController
         $junk->exchangeArray($data);
         $id = $this->getJunkFoodTable()->saveJunkFood($junk);
 
+        if (isset($data['ingredients']) && !empty($data['ingredients'])) {
+            $relations = array();
+
+            foreach ($data['ingredients'] as $ingredient) {
+                $relations = array('ingrID' => $ingredient['ingrID'], 'junkfoodID' => $id, 'gramm' => $ingredient['gramm']);
+            }
+
+            $this->getRelationTable()->saveRelation($relations);
+        }
+
         return new JsonModel(array(
-            'data' => 'Successfully saved with ID: ' . $id,
+            'status' => 'success',
+            'message' => 'Junkfood successfully created.'
         ));
     }
 
-    public function update($id, $data)
+    public function update($data)
     {
-        $data['id'] = $id;
-        $junk = $this->getJunkFoodTable()->getJunkFood($id);
+        if ($this->getAuthService()->hasIdentity()){
+            $this->userName = $this->getAuthService()->getIdentity();
+            $userId = $this->getUserTable()->getUserByName($this->userName);
+        } else {
+            return new JsonModel(array(
+                'status' => 'failure',
+                'message' => 'You have to log in first'
+            ));
+        }
+
+        $junk = new JunkFood();
+        $junk->exchangeArray($data);
+
+        if ($junk->userID == $userId) {
+            $this->getJunkFoodTable()->deleteJunkFood($id);
+
+            return new JsonModel(array(
+                'status' => 'success',
+                'message' => "Junkfood {$junk->name} successfully deleted!"
+            ));
+        }
+
         $id = $this->getJunkFoodTable()->saveJunkFood($junk);
 
+        if (isset($data['ingredients']) && !empty($data['ingredients'])) {
+            $relations = array();
+
+            foreach ($data['ingredients'] as $ingredient) {
+                $relations = array('ingrID' => $ingredient['ingrID'], 'junkfoodID' => $id, 'gramm' => $ingredient['gramm']);
+            }
+
+            $this->getRelationTable()->saveRelation($relations);
+        }
+
         return new JsonModel(array(
-            'data' => 'Successfully updated ' . $junk->type . ' with ID: ' . $id,
+            'status' => 'success',
+            'message' => "Junkfood {$junk->name} successfully updated."
         ));
     }
 
     public function delete($id)
     {
-        $this->getJunkFoodTable()->deleteJunkFood($id);
+        if ($this->getAuthService()->hasIdentity()){
+            $this->userName = $this->getAuthService()->getIdentity();
+            $userId = $this->getUserTable()->getUserByName($this->userName);
+        } else {
+            return new JsonModel(array(
+                'status' => 'failure',
+                'message' => 'You have to log in first'
+            ));
+        }
+
+        $junk = $this->getJunkFoodTable()->getJunkFood($id);
+
+        if ($junk->userID == $userId) {
+            $this->getJunkFoodTable()->deleteJunkFood($id);
+
+            return new JsonModel(array(
+                'status' => 'success',
+                'message' => "Junkfood {$junk->name} successfully deleted!"
+            ));
+        }
 
         return new JsonModel(array(
-            'data' => 'deleted',
+            'status' => 'failure',
+            'message' => "You may only delete your own junk!"
         ));
     }
 
@@ -127,5 +197,14 @@ class JunkFoodRestController extends AbstractRestfulController
             $this->userTable = $sm->get('Jfa\Model\UserTable');
         }
         return $this->userTable;
+    }
+
+    public function getRelationTable()
+    {
+        if (!$this->relationTable) {
+            $sm = $this->getServiceLocator();
+            $this->relationTable = $sm->get('Jfa\Model\JunkFoodIngredientTable');
+        }
+        return $this->relationTable;
     }
 }
